requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
openai==2.6.1
python-dotenv==1.2.1